requirements.txt
flask>=2.3,<4.0
werkzeug>=2.3,<4.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
